+++ R/nestcv.train.R
@@ -15,6 +15,11 @@
 #'   character vector with names of filtered predictors.
 #' @param filter_options List of additional arguments passed to the filter
 #'   function specified by `filterFUN`.
+#' @param balance Specifies method for dealing with imbalanced class data.
+#'   Current options are `"randomsample"` or `"smote"`. See [randomsample()] and
+#'   [smote()]
+#' @param balance_options List of additional arguments passed to the balancing
+#'   function
 #' @param outer_method String of either `"cv"` or `"LOOCV"` specifying whether
 #'   to do k-fold CV or leave one out CV (LOOCV) for the outer folds
 #' @param n_outer_folds Number of outer CV folds
@@ -41,7 +46,7 @@
 #'   if there are `NA` in 'y', but columns (predictors) containing `NA` are
 #'   removed from 'x' to preserve cases. Any other value means that `NA` are
 #'   ignored (a message is given).
-#' @param ... Arguments passed to [caret::train]
+#' @param ... Arguments passed to [caret::train] including `method`
 #' @return An object with S3 class "nestcv.train"
 #'   \item{call}{the matched call}
 #'   \item{output}{Predictions on the left-out outer folds}
@@ -52,6 +57,7 @@
 #'   \item{outer_folds}{List of indices of outer test folds}
 #'   \item{final_fit}{Final fitted caret model using best tune parameters}
 #'   \item{final_vars}{Column names of filtered predictors entering final model}
+#'   \item{summary_vars}{Summary statistics of filtered predictors}
 #'   \item{roc}{ROC AUC for binary classification where available.}
 #'   \item{trControl}{`caret::trainControl` object used for inner CV}
 #'   \item{bestTunes}{best tuned parameters from each outer fold}
@@ -125,6 +131,8 @@
 nestcv.train <- function(y, x,
                          filterFUN = NULL,
                          filter_options = NULL,
+                         balance = NULL,
+                         balance_options = NULL,
                          outer_method = c("cv", "LOOCV"),
                          n_outer_folds = 10,
                          outer_folds = NULL,
@@ -137,6 +145,8 @@ nestcv.train <- function(y, x,
                          ...) {
   nestcv.call <- match.call(expand.dots = TRUE)
   outer_method <- match.arg(outer_method)
+  if (!is.null(balance) & is.numeric(y)) {
+    stop("`balance` can only be used for classification")}
   ok <- checkxy(y, x, na.option)
   y <- y[ok$r]
   x <- x[ok$r, ok$c]
@@ -165,12 +175,14 @@ nestcv.train <- function(y, x,
     cl <- makeCluster(cv.cores)
     clusterExport(cl, varlist = c("outer_folds", "y", "x", 
                                   "filterFUN", "filter_options",
+                                  "balance", "balance_options",
                                   "metric", "trControl", "tuneGrid",
                                   "nestcv.trainCore", ...),
                   envir = environment())
     outer_res <- parLapply(cl = cl, outer_folds, function(test) {
       nestcv.trainCore(test, y, x,
                        filterFUN, filter_options,
+                       balance, balance_options,
                        metric, trControl, tuneGrid, ...)
     })
     stopCluster(cl)
@@ -178,6 +190,7 @@ nestcv.train <- function(y, x,
     outer_res <- mclapply(outer_folds, function(test) {
       nestcv.trainCore(test, y, x,
                        filterFUN, filter_options,
+                       balance, balance_options,
                        metric, trControl, tuneGrid, ...)
     }, mc.cores = cv.cores)
   }
@@ -215,6 +228,7 @@ nestcv.train <- function(y, x,
               outer_method = outer_method,
               outer_folds = outer_folds,
               dimx = dim(x),
+              y = y,
               final_fit = final_fit,
               final_vars = colnames(filtx),
               roc = caret.roc,
@@ -229,28 +243,30 @@ nestcv.train <- function(y, x,
 
 nestcv.trainCore <- function(test, y, x,
                              filterFUN, filter_options,
+                             balance, balance_options,
                              metric, trControl, tuneGrid, ...) {
-  filtx <- if (is.null(filterFUN)) x else {
-    args <- list(y = y[-test], x = x[-test, ])
-    args <- append(args, filter_options)
-    fset <- do.call(filterFUN, args)
-    x[, fset]
-  }
-  fit <- caret::train(x = filtx[-test, ], y = y[-test],
+  dat <- nest_filt_bal(test, y, x, filterFUN, filter_options,
+                       balance, balance_options)
+  ytrain <- dat$ytrain
+  ytest <- dat$ytest
+  filt_xtrain <- dat$filt_xtrain
+  filt_xtest <- dat$filt_xtest
+  
+  fit <- caret::train(x = filt_xtrain, y = ytrain,
                       metric = metric,
                       trControl = trControl,
                       tuneGrid = tuneGrid, ...)
-  predy <- predict(fit, newdata = filtx[test, ])
-  preds <- data.frame(predy=predy, testy=y[test])
+  predy <- predict(fit, newdata = filt_xtest)
+  preds <- data.frame(predy=predy, testy=ytest)
   if (is.factor(y)) {
-    predyp <- predict(fit, newdata = filtx[test, ], type = "prob")
+    predyp <- predict(fit, newdata = filt_xtest, type = "prob")
     # note predyp has 2 columns
     preds$predyp <- predyp[,2]
   }
-  rownames(preds) <- rownames(x)[test]
+  rownames(preds) <- rownames(filt_xtest)
   ret <- list(preds = preds,
               fit = fit,
-              nfilter = ncol(filtx))
+              nfilter = ncol(filt_xtrain))
   ret
 }
 
@@ -267,7 +283,13 @@ summary.nestcv.train <- function(object,
                              LOOCV = "leave-one-out CV\n"))
   cat("Inner loop: ", paste0(object$trControl$number, "-fold ",
                              object$trControl$method, "\n"))
-  cat(object$dimx[1], "observations,", object$dimx[2], "predictors\n\n")
+  balance <- object$call$balance
+  if (!is.null(balance)) {
+    cat("Balancing: ", balance, "\n")
+  }
+  cat(object$dimx[1], "observations,", object$dimx[2], "predictors\n")
+  if (!is.numeric(object$y)) print(c(table(object$y)))
+  cat("\n")
   nfilter <- unlist(lapply(object$outer_result, '[[', 'nfilter'))
   foldres <- object$bestTunes
   foldres$n.filter <- nfilter
@@ -289,3 +311,39 @@ predict.nestcv.train <- function(object, newdata, ...) {
     stop("newdata is missing some predictors", call. = FALSE)
   predict(object$final_fit, newdata = newdata[, object$final_vars], ...)
 }
+
+
+#' Summarise variables
+#' 
+#' @param x Matrix or dataframe with variables in columns
+#' @return A matrix with variables in rows and mean, median and SD for each
+#'   variable or number of levels if the variable is a factor. If `NA` are
+#'   detected, an extra column `n.NA` is added with the numbers of `NA` for each
+#'   variable.
+#' @importFrom Rfast colMedians colVars
+#' @export
+#' 
+summary_vars <- function(x) {
+  if (is.matrix(x)) {
+    if (!is.numeric(x)) return("Not numeric")
+    mat <- x
+    selCols <- TRUE
+  } else {
+    selCols <- unlist(lapply(x, is.numeric))
+    mat <- as.matrix(x[, selCols])
+  }
+  summary_mat <- cbind(colMeans(mat, na.rm = TRUE), colMedians(mat, na.rm = TRUE),
+                       colVars(mat, std = TRUE, na.rm = TRUE))
+  colnames(summary_mat) <- c("mean", "median", "sd")
+  if (any(!selCols)) {
+    nlevels_mat <- unlist(lapply(x[, !selCols], function(i) nlevels(as.factor(i))))
+    out <- matrix(NA, nrow = ncol(x), ncol = 4,
+                  dimnames = list(colnames(x), c("mean", "median", "sd", "nlevels")))
+    out[selCols, 1:3] <- summary_mat
+    out[!selCols, 4] <- nlevels_mat
+  } else out <- summary_mat
+  n.NA <- apply(x, 2, function(i) sum(is.na(i)))
+  if (any(n.NA > 0)) out <- cbind(out, n.NA)
+  out
+}
+